monster情報のpropsの受け渡し
