+++ src/components/layouts/index.js
@@ -4,6 +4,7 @@ import Headers from "./headers";
 import Footers from "./footers";
 import Loading from "./loading";
 import SecondaryButton from "./secondaryButton";
+import BlackButton from "./blackButton";
 
 export {
   BasicButton,
@@ -12,4 +13,5 @@ export {
   Footers,
   Loading,
   SecondaryButton,
+  BlackButton,
 };
